@@ -22,13 +22,13 @@
         "sort-packages": true
     },
     "require": {
-        "php": "^7.4 || ~8.0.0 || ~8.1.0",
+        "php": "~8.1.0 || ~8.2.0",
         "laminas/laminas-escaper": "^2.9",
         "laminas/laminas-stdlib": "^3.6"
     },
     "require-dev": {
         "laminas/laminas-coding-standard": "~2.3",
-        "laminas/laminas-servicemanager": "^3.8",
+        "laminas/laminas-servicemanager": "^3.21.0",
         "phpunit/phpunit": "^9.5.5"
     },
     "suggest": {
